Inserted the ScrollView of choices into a constraint layout for fling action to work under them
The empty space between the choices and buttons (if any) wasn't responding to fling action to change questions as
this was being filled by the ScrollView.

The ScrollView is necessary because, if the options are very long strings, and the screen size is small,
the user should be able to scroll to see all the options.

Why can't make it work with just the ScrollView?
Just removing ScrollView's bottom constraint and making ScrollView's height to wrap_content is giving the bug that
it is overlapping with buttons when the strings in the options are too long, in other words, as it's lower boundary
is not set, it is just going all the way down.

So, moved this ScrollView into a newly created ConstraintLayout where
 the constrainLayout fills all space and bounds the ScrollView
 the ScrollViews's height is set to wrap_content

Note: For testing this, one of the questions' options are made very long which is not committed into git.

Note: Android complained that ScrollView should not be used there and instead to make it NestedScrollView which I did.

On branch swipe-gestures
Changes to be committed:
 modified:   app/src/main/res/layout/activity_learn.xml