Converted UseCheckButton setting to radio button along with its related setting
The reason is that if the other setting is activated, Check button will not be of use as validation
is done right after selection of the choice

Consequently, a no-longer-needed data member is removed from SettingsActivity.

Note:
Even though now, it is just one setting, use_check_button_in_practice which takes true/false,
it is shown as a set of two radio radio buttons in the layout.
One makes the setting true and the other false. The reason is for the setting to be self-explanatory.

On branch development
Changes to be committed:
 modified:   app/src/main/java/com/ghsembrar/llrtestpreparation/SettingsActivity.java
 modified:   app/src/main/res/layout/activity_settings.xml